requirements.txt
discord.py
structlog
python-dotenv
requests
textdistance
arrow